+++ ajwah_block_examples/lib/main.dart
@@ -1,20 +1,18 @@
 import 'package:ajwah_bloc/ajwah_bloc.dart';
 import 'package:ajwah_block_examples/appStateProvider.dart';
 import 'package:ajwah_block_examples/counter/components/counterComponent.dart';
-import 'package:ajwah_block_examples/counter/store/counterEffect.dart';
-import 'package:ajwah_block_examples/counter/store/counterState.dart';
+import 'package:ajwah_block_examples/store/effects/searchEffect.dart';
+import 'package:ajwah_block_examples/store/states/SearchState.dart';
+import 'package:ajwah_block_examples/store/states/TodoState.dart';
+import 'package:ajwah_block_examples/store/states/counterState.dart';
 import 'package:ajwah_block_examples/todo/components/todoContainer.dart';
-import 'package:ajwah_block_examples/todo/store/TodoEffects.dart';
-import 'package:ajwah_block_examples/todo/store/TodoState.dart';
 import 'package:ajwah_block_examples/wikiSearch/components/searchComponent.dart';
-import 'package:ajwah_block_examples/wikiSearch/store/SearchState.dart';
-import 'package:ajwah_block_examples/wikiSearch/store/searchEffect.dart';
 import 'package:flutter_web/material.dart';
 
 void main() => runApp(App(
     store: createStore(
         states: [CounterState(), SearchState(), TodoState()],
-        effects: [CounterEffect(), SearchEffect(), TodoEffects()])));
+        effects: [SearchEffect()])));
 
 class App extends StatefulWidget {
   App({Key key, this.store}) : super(key: key);