+++ ajwah_block_examples/lib/widgets/popupMenu.dart
@@ -6,9 +6,9 @@ class PopupMemu extends StatelessWidget {
   const PopupMemu({Key key}) : super(key: key);
 
   menuSelect(BuildContext context, String path) {
-    if (path == '/todo') {
-      dispatch(ActionTypes.LoadingTodos);
-    }
+    // if (path == '/todo') {
+    //   dispatch(ActionTypes.LoadingTodos);
+    // }
     Navigator.pushReplacementNamed(context, path);
   }
 
@@ -19,25 +19,25 @@ class PopupMemu extends StatelessWidget {
         menuSelect(context, value);
       },
       itemBuilder: (ctx) => [
-            PopupMenuItem(
-              value: '/',
-              child: ListTile(
-                title: Text('Counter'),
-              ),
-            ),
-            PopupMenuItem(
-              value: '/search',
-              child: ListTile(
-                title: Text('Search'),
-              ),
-            ),
-            PopupMenuItem(
-              value: '/todo',
-              child: ListTile(
-                title: Text('Todo'),
-              ),
-            )
-          ],
+        PopupMenuItem(
+          value: '/',
+          child: ListTile(
+            title: Text('Counter'),
+          ),
+        ),
+        PopupMenuItem(
+          value: '/search',
+          child: ListTile(
+            title: Text('Search'),
+          ),
+        ),
+        PopupMenuItem(
+          value: '/todo',
+          child: ListTile(
+            title: Text('Todo'),
+          ),
+        )
+      ],
     );
   }
 }